crypto_app/requirements.txt
coinbase==2.1.0
awscrt==0.13.9
awsiotsdk==1.11.0